+++ test/bones_api_test_utils_test.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.14
+// BUILDER: reflection_factory/1.2.15
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -36,7 +36,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule> {
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.14');
+  Version get reflectionFactoryVersion => Version.parse('1.2.15');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>